requirements.txt
fandanGO-core
fandanGO-aria
python-dotenv>=1.0.1
requests>=2.31.0
tabulate>=0.9.0
